requirements.txt
# ── Core ──
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1

# ── Database ──
psycopg2-binary==2.9.11
dj-database-url==3.1.2

# ── Server ──
gunicorn==25.3.0
whitenoise==6.12.0

# ── CORS ──
django-cors-headers==4.9.0

# ── Cache ──
django-redis==6.0.0
redis==7.4.0
hiredis==3.3.1

# ── Email ──
django-simple-email-confirmation==0.71
sendgrid==6.12.5
python-http-client==3.3.7

# ── Files / Images ──
pillow==12.1.1
cloudinary==1.41.0
django-cloudinary-storage==0.3.0

# ── Config ──
python-decouple==3.8
python-dotenv==1.2.2

# ── Security / SSL ──
certifi==2026.2.25
cryptography==46.0.7
cffi==2.0.0
pycparser==3.0

# ── Misc ──
asgiref==3.11.1
sqlparse==0.5.5
six==1.17.0
packaging==26.0
MarkupSafe==3.0.3
Werkzeug==3.1.8
